booking detail page-top/layout
