+++ src/pages/booking/Left.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+function Left() {
+	return (
+		<>
+			<div className='BKItem'>left</div>
+		</>
+	);
+}
+
+export default Left;
