+++ src/pages/booking/Right.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+function Right() {
+	return (
+		<>
+			<div className='BKItem'>right</div>
+		</>
+	);
+}
+
+export default Right;
