+++ src/pages/roomsdetail/DeatilBooking.js
@@ -484,7 +484,7 @@ function DeatilBooking(props) {
 						if (token) {
 							if (stime !== '' || etime !== '') {
 								window.location.href =
-									'/booking/detail?num=' +
+									'/booking/main?num=' +
 									num +
 									'&date=' +
 									moment(selectDay).format('YYYY-MM-DD') +
